[SC-26] multiple product versions (#6)
This is a refactor to support multiple product versions. This
implementation should work however it doesn't due to an issue
with how CloudformationProduct work.

We have sent AWS a ticket regarding this issue and are waiting
to see if they can provide any assistance.  We are currently
blocked so wanted to PR this and save it until get a
reply from AWS.

Summary of AWS issue:
The cloudformation AWS::ServiceCatalog::CloudFormationProduct
Property ProvisioningArtifactParameters is not working correctly.
It duplicates the entire list of products on an update.  This
only happens after a service catalog action is associated with
one of the Product version (i.e. ProvisioningArtifactIds).
This is bad because you can wind up with a ton of product versions
that do not match what's in the cloudformation template.